Added Microsoft Defender compliance check and Azure Function trigger
This commit includes the following key components:

main.zsh: A Zsh script that checks the status of Microsoft Defender on a macOS endpoint. If Microsoft Defender is not running or uninstalled, the script triggers an Azure Function for remediation actions. The script runs every 15 minutes.

com.macOSComplianceChecker.plist: A plist file that schedules the execution of main.zsh every 15 minutes.

README.md: Updated with sections on project introduction, installation guide, usage, contributing, testing, and license information.

These components form the core of our Microsoft Defender compliance checking mechanism, ensuring that remediation actions are taken when non-compliance is detected.

Still need to do a guide on usage and where the files need to be stored etc...